% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_likelihood.r
\name{context_likelihood}
\alias{context_likelihood}
\title{context_likelihood}
\usage{
context_likelihood(x, batch, p0 = 0.2)
}
\arguments{
\item{x}{a gene expression matrix (normalized and scaled)}

\item{batch}{a numeric vector indicating the batch labels of each cell}

\item{p0}{cutoff for context likelihood neighbors (default: 0.2)}
}
\value{
a matrix of context likelihood neighbors (CLN)
}
\description{
Find the mutually similar cells between batches using the context likelihood
}
\author{
Wuming Gong
}
