% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{simulateTemporalRNAseq}
\alias{simulateTemporalRNAseq}
\title{simulateTemporalRNAseq}
\usage{
simulateTemporalRNAseq(ncells, time.points, K = 5, batch.effect = 1,
  N = 2000)
}
\arguments{
\item{ncells}{a matrix of number of cells.  The number of columns represents the batches and the number of rows represents the lineages (H).}

\item{time.points}{a matrix of distribution of batches along the time indices.  The number of columns represents the batches and the rowsrepresents time points(T).}

\item{K}{the number of dimensions of the low dimensional biological subspace.}

\item{batch.effect}{the strength of the added batch effect, which corresponds to the weight of the Gaussian random batch vector.}
}
\value{
A SummarizedExperiment object of simulated temporal scRNA-seq data
}
\description{
Simulating time series single cell RNA-seq data with batch effects
}
\details{
The simulation of temporal scRNA-seq data follows three basic assumptions: 
(1) the developmental trajectories can be represented within a K-dimensional biological subspace.  
(2) We assumed that all H lineages arise from the same progenitors and differentiated toward different directions.  
Thus, to simulate the developmental trajectory of any lineage h∈{1,⋯,H} , we use the origin of the K-dimensional space as the starting point, randomly select another point in the K-dimensional space as the terminal point, and draw a lineage segment (a line segment) that connects the origin and the terminal point.  
This lineage segment therefore represents the developmental trajectory of lineage h on the K-dimensional biological subspace.  
(3) The developmental speed is constant.  Thus, the each lineage segment is evenly split into T parts, where each part is associate with the time index from 1 to T.  
To simulate the cells of lineage h from time t∈{1,⋯,T}, we randomly draw points from the lineage segments that correspond to lineage h from time t on the K-dimensional space.  Note that for balanced data, the points are drawn from the entire lineage segments, while for the imbalanced data, the points can only be drawn from the lineage segments covered by the corresponding time period.  Then to simulate high-dimensional gene expression, the K-dimensional points are projected to N-dimensional space by a randomly Gaussian matrix.  Similar to Haghverdi et al., the batch effects are incorporated by generating a Gaussian random vector for each batch and adding it to the gene expression matrix12.
}
\examples{
library(scNCA)
# Let us first simulate a balanced temporal scRNA-seq data with 2,000 genes, three lineages, 
# three batches and five time points. In the balanced data, each batch covers the cells from all lineages and time points. 
# We assume that there are 100 cells for each lineage/batch combination so there are in total 900 cells:
ncells <- cbind(
	c(100, 100, 100),   # batch i
 c(100, 100, 100),   # batch ii
 c(100, 100, 100)    # batch iii
)
time.points <- cbind(
	c(1, 1, 1, 1, 1),   # batch i
	c(1, 1, 1, 1, 1),   # batch ii
	c(1, 1, 1, 1, 1)    # batch iii
)
set.seed(1)
se <- simulateTemporalRNAseq(ncells = ncells, time.points = time.points)

}
\author{
Wuming Gong
}
