% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsBET.r
\name{tsBET}
\alias{tsBET}
\title{tsBET}
\usage{
tsBET(x, batch, time.table, resample = 100, neighborhood_size = 100,
  mc.cores = 8)
}
\arguments{
\item{x}{a M by M sparse matrix representing the adjacency graph of the cells}

\item{batch}{a numeric vector indicating the batch labels of each cell}

\item{time.table}{a binary matrix indicating the distribution of cells at each time point}

\item{resample}{number of resampling for building the null distribution of local batch score (default: 100)}

\item{neighborhood_size}{number of neighboring cells when computing the local batch score (default: 100)}

\item{mc.cores}{number of CPU cores (default: 8)}
}
\value{
a list of tsBET results
}
\description{
Testing for batch effets in temporal scRNA-seq data
batch effect test for temporal scRNA-seq data
}
\author{
Wuming Gong
}
